Don't install network-manager packages that call home.
Ubuntu has added a suggested package which installs a config file for network manager.
This config file enables a feature which checks the connection is actually connected
to the internet. It does this by connecting to a website on the internet every five
minutes. Due to privacy concerns, Lubuntu is NOT installing this package by default,
leaving it to the user to decide if he/she wants this feature.